updated queue version

+++ topological_sort/queue_sequential/c_queue_sequential_topological_sort.c
@@ -6,7 +6,7 @@
 extern void* asm_balloc(uint32_t size);
 extern void asm_balloc_free(void* memory_pointer);
 
-uint8_t c_algorithm_t_queue_sequential(uint8_t n, Pair* input_pairs, uint8_t input_pairs_len, uint32_t* output) {
+uint8_t c_algorithm_t_queue_sequential(uint32_t n, Pair* input_pairs, uint32_t input_pairs_len, uint32_t* output) {
 	// T1 (Initialize)
 	TopologicalNode* P;
 	uint32_t FRONT;
